+++ docs/examples/buttonlink/main.js
@@ -1,16 +1,16 @@
 // @flow strict
 import { type Node } from 'react';
-import { Button, Flex } from 'gestalt';
+import { ButtonLink, Flex } from 'gestalt';
 
 export default function Example(): Node {
   return (
     <Flex alignItems="center" height="100%" justifyContent="center" width="100%">
-      <Button
+      <ButtonLink
         accessibilityLabel="Visit Pinterest"
         iconEnd="visit"
         size="lg"
+        color="red"
         text="Visit Pinterest"
-        role="link"
         href="https://pinterest.com"
       />
     </Flex>